requirements.txt
Flask
flask-apispec==0.11.4
Flask-RESTful==0.3.9
Flask-Session==0.4.0
flask-cors
gunicorn
apispec==6.1.0
marshmallow==3.19.0
flask-apispec
pymongo
bcrypt
email-validator
werkzeug
pyvips
pyotp==2.7.0
requests==2.28.2
Pillow
jwt
